address review: RecentPricesRequest counts are u32, not i64
max/pageSize/pageNumber are positive counts interpolated into the query
string; unsigned makes a negative count unrepresentable at the type level.

+++ src/model/requests.rs
@@ -25,13 +25,13 @@ pub struct RecentPricesRequest<'a> {
     pub to: Option<&'a str>,
     /// Optional max number of price points (default: 10)
     #[serde(rename = "max")]
-    pub max_points: Option<i64>,
+    pub max_points: Option<u32>,
     /// Optional page size (default: 20, disable paging = 0)
     #[serde(rename = "pageSize")]
-    pub page_size: Option<i64>,
+    pub page_size: Option<u32>,
     /// Optional page number (default: 1)
     #[serde(rename = "pageNumber")]
-    pub page_number: Option<i64>,
+    pub page_number: Option<u32>,
 }
 
 impl<'a> RecentPricesRequest<'a> {
@@ -67,21 +67,21 @@ impl<'a> RecentPricesRequest<'a> {
 
     /// Set the max points
     #[must_use]
-    pub fn with_max_points(mut self, max_points: i64) -> Self {
+    pub fn with_max_points(mut self, max_points: u32) -> Self {
         self.max_points = Some(max_points);
         self
     }
 
     /// Set the page size
     #[must_use]
-    pub fn with_page_size(mut self, page_size: i64) -> Self {
+    pub fn with_page_size(mut self, page_size: u32) -> Self {
         self.page_size = Some(page_size);
         self
     }
 
     /// Set the page number
     #[must_use]
-    pub fn with_page_number(mut self, page_number: i64) -> Self {
+    pub fn with_page_number(mut self, page_number: u32) -> Self {
         self.page_number = Some(page_number);
         self
     }
